move atlas_sql_test_connection back to core
